Bugfix - added #ifdef checks around SPI and ADC entries in dmaoptEntryTable in cli.c
Details: if you didn't have USE_ADC or USE_SPI defined, the ADC and SPI entries remained defined in dmaoptEntryTable.
When you switch to the CLI tab in the Configurator, it immediatelly calls the dump command, to build its AutoComplete
cache, this calls printDmaopt and it attempts to iterate thru the entries in dmaoptEntryTable and crashes Betaflight
on trying to check the ADC and SPI entries.